refactor: schema.js into resolvers and mutations file
